+++ Source/Services/VitalRustService/src/main.rs
@@ -2,25 +2,17 @@
     all(not(debug_assertions), target_os = "windows"),
     windows_subsystem = "windows"
 )]
-use std::collections::{HashMap, HashSet};
-use std::string;
 use std::time::Duration;
 use std::{thread, time::Instant};
 extern crate nvml_wrapper as nvml;
-use chrono::{DateTime, Utc};
-use generated_vital_rust_service_api_def::{
-    DiskLoad, DiskThroughput, NetworkAdapterProperties, NetworkAdapterUsage, NetworkAdapterUtil,
-    ProcessGpuUtil,
-};
 use log::{error, info, LevelFilter};
 use log::{Level, Metadata, Record};
 
 use crate::commands::get_vital_service_ports;
-use crate::generated_vital_rust_service_api_def::{PidProcessTitleMapping, SendUtilizationRequest};
-use nvml::{struct_wrappers::device::ProcessUtilizationSample, NVML};
-use sysinfo::{
-    ComponentExt, DiskExt, DiskType, NetworkExt, PidExt, Process, ProcessorExt, SystemExt,
-};
+use crate::generated_vital_rust_service_api_def::SendUtilizationRequest;
+use crate::software::get_process_util;
+use nvml::NVML;
+use sysinfo::SystemExt;
 use systemstat::Platform;
 use tokio::join;
 mod api;
@@ -29,9 +21,9 @@ mod generated_client_api_dto_def;
 mod generated_vital_rust_service_api_def;
 mod machine;
 mod nvidia;
-mod windows;
+pub mod software;
+
 use api::post_request;
-use sysinfo::ProcessExt;
 static LOGGER: SimpleLogger = SimpleLogger;
 static WAIT_TIME: core::time::Duration = Duration::from_millis(1000);
 
@@ -98,80 +90,6 @@ async fn main() {
         }
         info!("time taken: {}", now.elapsed().as_millis());
     }
-}
-
-fn get_process_util(
-    sysinfo: &sysinfo::System,
-    nvml: &Option<NVML>,
-    time_stamp: DateTime<Utc>,
-) -> Option<Vec<generated_vital_rust_service_api_def::ProcessData>> {
-    let mut list = Vec::new();
-    let processes = sysinfo.processes();
-
-    let process_gpu_utilization_samples = nvidia::get_process_gpu_util(&nvml).unwrap_or(Vec::new());
-    /*
-    let using_compute = gpu_device.running_compute_processes().unwrap();
-    let using_graphics = gpu_device.running_graphics_processes().unwrap();
-
-    for p in using_compute {
-        p.used_gpu_memory();
-    } */
-
-    let cores = sysinfo.physical_core_count();
-    let main_window_titles = windows::get_mainwindowtitles();
-    for (pid, process) in processes {
-        let disk_bytes = process.disk_usage();
-        // get first gpu usage that has this pid
-
-        let pid = pid.as_u32();
-        let path = windows::get_process_Path(pid); // takes some time
-
-        let mut description: Option<String> = None; // takes some time
-                                                    /*         if path.is_some() {
-                                                        description = match windows::get_file_description(path.to_owned().unwrap().to_string())
-                                                        {
-                                                            Ok(title) => Some(title),
-                                                            Err(_) => None,
-                                                        };
-                                                    }; */
-        // windows::get_process_ideal_processors(pid); //takes a lot of time
-        list.push(generated_vital_rust_service_api_def::ProcessData {
-            name: process.name().to_string(),
-            pid: pid as f64,
-            main_window_title: match main_window_titles.get(&pid) {
-                Some(title) => Some(title.to_string()),
-                None => None,
-            },
-            description: description,
-            executable_path: path,
-            parent_pid: match process.parent() {
-                Some(pid) => Some(pid.as_u32() as f64),
-                None => None,
-            },
-            cpu_percentage: (process.cpu_usage() / cores.unwrap() as f32) as f64,
-            memory_kb: process.memory() as f64,
-            disk_usage: generated_vital_rust_service_api_def::ProcessDiskUsage {
-                read_bytes_per_second: disk_bytes.read_bytes as f64,
-                write_bytes_per_second: disk_bytes.written_bytes as f64,
-            },
-            status: Some(process.status().to_string()),
-            gpu_util: match process_gpu_utilization_samples
-                .iter()
-                .find(|sample| sample.pid == pid)
-            {
-                Some(util) => Some(ProcessGpuUtil {
-                    gpu_core_percentage: Some(util.sm_util as f64),
-                    gpu_decoding_percentage: Some(util.dec_util as f64),
-                    gpu_encoding_percentage: Some(util.enc_util as f64),
-                    gpu_mem_percentage: Some(util.mem_util as f64),
-                }),
-                None => None,
-            },
-
-            time_stamp: time_stamp.to_rfc3339(),
-        });
-    }
-    return Some(list);
 }
 
 struct SimpleLogger;
